Remove namespace from helpers
