+++ app/helpers/people_helper.rb
@@ -1,12 +1,10 @@
-require 'peoplefinder'
-
-module Peoplefinder::PeopleHelper
+module PeopleHelper
   def day_name(symbol)
-    I18n.t(symbol, scope: [:peoplefinder, :people, :day_names])
+    I18n.t(symbol, scope: [:people, :day_names])
   end
 
   def day_symbol(symbol)
-    I18n.t(symbol, scope: [:peoplefinder, :people, :day_symbols])
+    I18n.t(symbol, scope: [:people, :day_symbols])
   end
 
   def profile_image_tag(person, source, options = {})
