+++ app/helpers/versions_helper.rb
@@ -1,12 +1,10 @@
-require 'peoplefinder'
-
-module Peoplefinder::VersionsHelper
+module VersionsHelper
   def view_template(version)
     version.membership? ? 'membership' : 'general'
   end
 
   def link_to_edited_item(version)
-    if version.item && !version.item.is_a?(Peoplefinder::Membership)
+    if version.item && !version.item.is_a?(Membership)
       link_to(version.item, version.item)
     end
   end
